Extract ErrorDescription into separate module
